finagle-http: Disable DeadlineFilter in HTTP servers
Problem

We found out DeadlineFilter and BadRequestHandler are created for every Http2
stream.  Their field members are also created at the same frequency, such as
counters. This waste some resources.  In addition, DeadlineFilter is not working
in the H2 server stack because it relies on a connection shared leaky bucket to
reject, but the leaky bucket is created per stream.

Solution

Disable DeadlineFilter for H2 server.
Make BadRequestHandler sharable. It needs the stack param as a parameter and
then pass to all pipelines. Make the initServer partially applied so it is
actually shared.

Result

We see H2 server throughput increased with the change.

Differential Revision: https://phabricator.twitter.biz/D964479