[util] Submit continuations separately to the scheduler in Promise
Problem:
We submit a Promise's entire WaitQueue to the scheduler.

Solution:
Submit each continuation to the scheduler on its own. This could have subtle changes to the order at which continuations are executed, but we make no guarantees about ordering.

This opens us up to further scheduler submission changes that assume continuations are submitted one at a time.

JIRA Issues: CSL-12265

Differential Revision: https://phabricator.twitter.biz/D973092