[util-core] Move scheduler submission into Fiber stored in Local
Problem:
We want to introduce Fibers.

Solution:
Introduce the default Fiber implementation and hand off all scheduler submission
to Fibers. The fibers themselves will be stored in the Local context. For now,
there is no behavior change because the default Fiber only submits tasks to the
Scheduler.

JIRA Issues: CSL-12266

Differential Revision: https://phabricator.twitter.biz/D976677