util-stats: Add notion of an undetermined identity type
Problem

We want to be able to exclude some metrics from being dimensional
because they will become redundant. However, we don't have a way to
express that right now.

Solution

Add an `NonDeterminate` state for our metric identity. This becomes the
default state. As metric names this flag can be refined to express that
a metric name is only valid hierarchically.

JIRA Issues: CSL-12091

Differential Revision: https://phabricator.twitter.biz/D964827