util: Fix flaky FuturePool ResourceTracker test
Problem

The ResourceTracker test in FuturePoolTest is flaky, with the tracker
occasionally registering an extra update. This is caused by the `Await.result`
call in the FuturePool thread occasionally adding a continuation to the
Promise, which also has the resource tracker present, to tell when the task
is complete.

Solution

Switch to using a CountdownLatch to (potentially) pause the FuturePool thread.

JIRA Issues: CSL-12352

Differential Revision: https://phabricator.twitter.biz/D984397