[util-core] Take care of thread flushing in Fiber
Problem:
When we start using the work queue based fiber we're going to need a way to call flush() for its tasks.

Solution:
Transfer the responsibility for flushing to the local fiber. This fiber will perform any necessary flushing of its own queue and the underlying scheduler. For example:
- in the default case, Fiber.Global, only the global scheduler needs to be flushed
- in the case of a WorkQueueFiber backed by a thread local scheduler, we need to flush both the fiber's thread local queue and the current thread's queue in the LocalScheduler

JIRA Issues: CSL-12359

Differential Revision: https://phabricator.twitter.biz/D991279