[util-core] Add resource tracking to WorkQueueFiber
Problem:
We want to do CPU resource tracking in fibers.

Solution:
Add a resource tracker to the WorkQueueFiber and track the cpu time to run the entire work loop. If the fiber is flushed it may do extra work, so subtract the cpu time it takes to flush.

JIRA Issues: CSL-12268

Differential Revision: https://phabricator.twitter.biz/D1007425