util/util-app: Include the exception in the error message printed when the application encounters an error on startup
Problem

When the application exits prematurely due to an exception encountered on startup, this
exception is not printed to stderr, only the message "Exception thrown in main on startup".
The exception stack trace is printed to stdout, but this is unintuitive for the user.

Solution

Include the exception and stack trace in the exception printed to stderr.

Differential Revision: https://phabricator.twitter.biz/D1116753