util-hashing: Faster ConsistentHashingDistributor
Problem
ConsistentHashingDistributor uses TreeMap to find a node for a key. TreeMap is relatively slow.

Solution
Replace the TreeMap with two arrays, one sorted for keys and one matching the first for values. Use binarySearch to find the matching key.
Because keys stored in array are not scattered in memory and are primitive values binary search is much faster. Based on ad-hoc benchmarking
the gain is at least 50%

Differential Revision: https://phabricator.twitter.biz/D1175980