util/util-core: Introduce ImmediateValueFuture (continuations execute immediately)
Problem

The transforming continuations (ie not side-effects) of `Future.const` go
through the scheduler. While this is safe in all use cases, it can have a
negative impact on execution times under certain scenarios due to
reordering.

Solution

Introduce `ImmediateValueFuture` that bypasses the scheduler in all cases
and executes continuations immediately. This Future type is not safe to use
in all scenarios and we outline this in the doc.

Differential Revision: https://phabricator.twitter.biz/D1285047